Update ISO27001 annex control to match latest revision
The 2022 revision of the ISO27001 standard changed the numbering of the
Annex A controls - Secure development life cycle is now A.8.25

The 2022 revision has been out long enough that there should now be no
organisations implementing the 2013 revision; so it's reasonable to
assume that people know the new Annex A Controls numbering.

Note that this commit uses the same capitalisation as the standard
itself, with the words "development", "life" and "cycle" in lower case.
This doesn't match IEC62304 control on the next line of the readme, so
stylistically it might be jarring to some readers.